ft: design tooltip component - move tooltip component folder into the components folder - setup enzyme for tests
