--- src/tooltip/tooltip.scss
@@ -1,18 +1,0 @@
-.tool-tip-container {
-  position: relative;
-  display: inline-block;
-  border-bottom: 1px dotted black;
-  background-color: black;
-  color: #fff;
-  padding-left: 15px;
-  border-radius: 3px;
-
-  .tooltip-text {
-    padding-right: 15px;
-    cursor: pointer;
-    background: black;
-    color: white;
-    border-color: black;
-    outline: none;
-  }
-}
